requirement.txt
absl-py==2.1.0
addict==2.4.0
affine==2.4.0
aioboto3==13.2.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
autocommand==2.2.2
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.1.0
blinker==1.8.2
boto3==1.35.36
botocore==1.35.36
cachetools==5.5.0
causal-conv1d
ccimport==0.4.4
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
cumm-cu116==0.4.11
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
fire==0.7.0
flake8==7.1.1
Flask==3.0.3
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
geopandas
geos==0.2.3
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.4
imageio==2.36.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
llvmlite==0.41.1
lxml==5.3.0
lyft-dataset-sdk==0.0.8
mamba_ssm==1.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mmcls==0.25.0
mmdet==2.28.2
mmsegmentation==0.30.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.58.0
numpy==1.23.5
nuscenes-devkit==1.1.10
oauthlib==3.2.2
opencv-python==4.8.1.78
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pccm==0.4.16
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
plyfile==1.1
portalocker==2.10.1
prettytable==3.7.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyquaternion==0.9.9
pytest==8.3.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.3.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.3
Shapely==1.7.1
six==1.15.0
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.6
spconv-cu116==2.3.6
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.2
torch_scatter==2.1.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trimesh==2.35.39
triton==3.1.0
typeguard==4.1.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.9.0
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
yapf==0.33.0
yarl==1.15.5
zipp==3.19.2